코드리뷰: 파일 분리, 네이밍 수정
